examples/lock-app/esp32: update lock, not onoff
The existing code would present as a doorlock device, but would not
actually ever update the state of the doorlock.  Changed to the same
style used in most in other example platforms, and now it correctly
updates, when tested against homeassistant.  Tested with both local
button press, and matter controller actions.

Signed-off-by: Karl Palsson <karlp@tweak.au>